refactor: 어드민/사용자 예약 API 경로 분리 (/admin/reservations, /reservations)
