+++ src/test/java/roomescape/TimeTest.java
@@ -83,12 +83,12 @@ public class TimeTest {
     RestAssured.given().log().all()
         .contentType(ContentType.JSON)
         .body(reservation)
-        .when().post("/reservations")
+        .when().post("/admin/reservations")
         .then().log().all()
         .statusCode(201);
 
     RestAssured.given().log().all()
-        .when().get("/reservations")
+        .when().get("/admin/reservations")
         .then().log().all()
         .statusCode(200)
         .body("size()", is(1));
