+++ src/test/java/roomescape/reservation/controller/ReservationControllerTest.java
@@ -2,12 +2,13 @@ package roomescape.reservation.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.time.LocalDateTime;
 import java.util.List;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -16,14 +17,18 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
+
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.exception.DuplicateReservationException;
+import roomescape.reservation.exception.PastReservationException;
 import roomescape.reservation.exception.ReservationNotFoundException;
 import roomescape.reservation.service.ReservationService;
 import roomescape.theme.domain.Theme;
 
 @WebMvcTest(ReservationController.class)
-public class ReservationControllerTest {
+class ReservationControllerTest {
+
     @Autowired
     private MockMvc mockMvc;
 
@@ -31,17 +36,18 @@ public class ReservationControllerTest {
     private ReservationService reservationService;
 
     @Test
-    void 예약_목록_조회_테스트() throws Exception {
+    void 이름으로_예약_목록_조회_테스트() throws Exception {
         // given
         ReservationTime time = new ReservationTime(1L, LocalDateTime.of(2030, 6, 1, 10, 0), LocalDateTime.of(2030, 6, 1, 12, 0));
         Theme theme = new Theme("테마", "설명", "https://img.test/a.png").withId(1L);
         Reservation reservation = new Reservation("라이", time, 1L)
                 .withId(1L)
                 .withTheme(theme);
-        Mockito.when(reservationService.getAll()).thenReturn(List.of(reservation));
+        Mockito.when(reservationService.getByName("라이")).thenReturn(List.of(reservation));
 
         // when & then
         mockMvc.perform(get("/reservations")
+                        .param("name", "라이")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.length()").value(1))
@@ -50,89 +56,12 @@ public class ReservationControllerTest {
     }
 
     @Test
-    void 예약_생성_테스트() throws Exception {
-        // given
-        ReservationTime time = new ReservationTime(1L, LocalDateTime.of(2030, 6, 1, 10, 0), LocalDateTime.of(2030, 6, 1, 12, 0));
-        Theme theme = new Theme("테마", "설명", "https://img.test/a.png").withId(1L);
-        Reservation saved = new Reservation("라이", time, 1L)
-                .withId(1L)
-                .withTheme(theme);
-        Mockito.when(reservationService.create(Mockito.any())).thenReturn(saved);
-
-        String requestBody = """
-                {
-                    "name": "라이",
-                    "themeId": 1,
-                    "timeId": 1
-                }
-                """;
-
-        // when & then
-        mockMvc.perform(post("/reservations")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestBody))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.id").value(1L))
-                .andExpect(jsonPath("$.name").value("라이"));
-    }
-
-    @DisplayName("빈 문자열 이름으로 예약 요청인 경우, MethodArgumentNotValidException이 발생한다.")
-    @Test
-    void 빈_이름_예약_생성_400_반환_테스트() throws Exception {
-        String requestBody = """
-                {
-                    "name": "",
-                    "themeId": 1,
-                    "timeId": 1
-                }
-                """;
-
-        mockMvc.perform(post("/reservations")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestBody))
-                .andExpect(status().isBadRequest());
-    }
-
-    @DisplayName("공백으로만 이루어진 이름으로 예약 요청인 경우, MethodArgumentNotValidException이 발생한다.")
-    @Test
-    void 공백_이름_예약_생성_400_반환_테스트() throws Exception {
-        String requestBody = """
-                {
-                    "name": "   ",
-                    "themeId": 1,
-                    "timeId": 1
-                }
-                """;
-
-        mockMvc.perform(post("/reservations")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestBody))
-                .andExpect(status().isBadRequest());
-    }
-
-    @DisplayName("timeId 없이 예약 요청인 경우, MethodArgumentNotValidException이 발생한다.")
-    @Test
-    void timeId_누락_예약_생성_400_반환_테스트() throws Exception {
-        String requestBody = """
-                {
-                    "name": "라이",
-                    "themeId": 1
-                }
-                """;
-
-        mockMvc.perform(post("/reservations")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestBody))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
     void 예약_취소_테스트() throws Exception {
         // when & then
         mockMvc.perform(delete("/reservations/{id}", 1L))
                 .andExpect(status().isNoContent());
 
-        Mockito.verify(reservationService).cancel(1L);
+        Mockito.verify(reservationService).cancelForUser(1L);
     }
 
     @DisplayName("존재하지 않는 예약을 취소하는 경우, 404를 반환한다.")
@@ -140,10 +69,121 @@ public class ReservationControllerTest {
     void 존재하지_않는_예약_취소_404_반환_테스트() throws Exception {
         // given
         Mockito.doThrow(new ReservationNotFoundException(999L))
-                .when(reservationService).cancel(999L);
+                .when(reservationService).cancelForUser(999L);
 
         // when & then
         mockMvc.perform(delete("/reservations/{id}", 999L))
                 .andExpect(status().isNotFound());
+    }
+
+    @DisplayName("이미 지난 예약을 취소하는 경우, 400을 반환한다.")
+    @Test
+    void 지난_예약_취소_400_반환_테스트() throws Exception {
+        // given
+        Mockito.doThrow(PastReservationException.pastCancel())
+                .when(reservationService).cancelForUser(1L);
+
+        // when & then
+        mockMvc.perform(delete("/reservations/{id}", 1L))
+                .andExpect(status().isBadRequest());
+    }
+
+    @Test
+    void 예약_변경_테스트() throws Exception {
+        // given
+        ReservationTime newTime = new ReservationTime(2L, LocalDateTime.of(2030, 6, 1, 14, 0), LocalDateTime.of(2030, 6, 1, 16, 0));
+        Theme theme = new Theme("테마", "설명", "https://img.test/a.png").withId(1L);
+        Reservation updated = new Reservation("라이", newTime, 1L)
+                .withId(1L)
+                .withTheme(theme);
+        Mockito.when(reservationService.update(1L, 2L)).thenReturn(updated);
+
+        String requestBody = """
+                {
+                    "timeId": 2
+                }
+                """;
+
+        // when & then
+        mockMvc.perform(put("/reservations/{id}", 1L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestBody))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.id").value(1L))
+                .andExpect(jsonPath("$.name").value("라이"));
+    }
+
+    @DisplayName("존재하지 않는 예약을 변경하는 경우, 404를 반환한다.")
+    @Test
+    void 존재하지_않는_예약_변경_404_반환_테스트() throws Exception {
+        // given
+        Mockito.when(reservationService.update(Mockito.anyLong(), Mockito.anyLong()))
+                .thenThrow(new ReservationNotFoundException(999L));
+
+        String requestBody = """
+                {
+                    "timeId": 2
+                }
+                """;
+
+        // when & then
+        mockMvc.perform(put("/reservations/{id}", 999L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestBody))
+                .andExpect(status().isNotFound());
+    }
+
+    @DisplayName("변경하려는 시간 슬롯이 과거인 경우, 400을 반환한다.")
+    @Test
+    void 과거_슬롯으로_예약_변경_400_반환_테스트() throws Exception {
+        // given
+        Mockito.when(reservationService.update(Mockito.anyLong(), Mockito.anyLong()))
+                .thenThrow(PastReservationException.pastReservation());
+
+        String requestBody = """
+                {
+                    "timeId": 1
+                }
+                """;
+
+        // when & then
+        mockMvc.perform(put("/reservations/{id}", 1L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestBody))
+                .andExpect(status().isBadRequest());
+    }
+
+    @DisplayName("변경하려는 시간이 이미 차 있는 경우, 409를 반환한다.")
+    @Test
+    void 이미_찬_시간으로_예약_변경_409_반환_테스트() throws Exception {
+        // given
+        Mockito.when(reservationService.update(Mockito.anyLong(), Mockito.anyLong()))
+                .thenThrow(new DuplicateReservationException());
+
+        String requestBody = """
+                {
+                    "timeId": 1
+                }
+                """;
+
+        // when & then
+        mockMvc.perform(put("/reservations/{id}", 1L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestBody))
+                .andExpect(status().isConflict());
+    }
+
+    @DisplayName("timeId 없이 변경 요청하는 경우, 400을 반환한다.")
+    @Test
+    void timeId_누락_예약_변경_400_반환_테스트() throws Exception {
+        String requestBody = """
+                {
+                }
+                """;
+
+        mockMvc.perform(put("/reservations/{id}", 1L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestBody))
+                .andExpect(status().isBadRequest());
     }
 }
